moved ineligible student method to IneligibleStudentService
